% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{budget_shares}
\alias{budget_shares}
\title{Budget shares}
\usage{
budget_shares(P, Q)
}
\arguments{
\item{P}{An m x n matrix of prices (row country, column goods).}

\item{Q}{An m x n matrix of quantities (row country, column goods).}
}
\value{
An m x n matrix of budget shares
}
\description{
Returns a matrix of budget shares corresponding to prices and quantities.
}
