% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ccd}
\alias{ccd}
\title{CCD (Caves, Christensen, and Diewert) index}
\usage{
ccd(P, Q)
}
\arguments{
\item{P}{An m x n matrix (or dataframe) of prices (row country, column goods).}

\item{Q}{An m x n matrix (or dataframe) of quantities (row country, column goods).}
}
\value{
An m-vector of income (relative to minimum income).
}
\description{
CCD (Caves, Christensen, and Diewert) index
}
\examples{
P <- matrix(1:6, ncol = 2)
Q <- matrix(2:7, ncol = n)
ccd(P, Q)
}
