% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{eks}
\alias{eks}
\title{EKS quantity index}
\usage{
eks(P, Q)
}
\arguments{
\item{P}{An m x n matrix (or dataframe) of prices (row country, column goods).}

\item{Q}{An m x n matrix (or dataframe) of quantities (row country, column goods).}
}
\value{
An m-vector of income (relative to minimum income).
}
\description{
EKS quantity index
}
\examples{
P <- matrix(1:6, ncol = 2)
Q <- matrix(2:7, ncol = n)
eks(P, Q)
}
