% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{gk}
\alias{gk}
\title{Gheary Khamis world prices and quantity index}
\usage{
gk(P, Q, pop = NULL)
}
\arguments{
\item{P}{An m x n matrix of prices (row country, column goods).}

\item{Q}{An m x n matrix of quantities (row country, column goods).}

\item{pop}{Optional m-vector of population numbers (if Q is per capita).}
}
\value{
A list with two elements: pi, an n-vector of prices and y, an m-vector of incomes.
}
\description{
Calculates the Gheary Khamis "world prices" and the corresponding
quantity index (normalized by the prices of the last good beting set
to unity).
}
\details{
For the Gheary Khamis index, it makes a difference whether the quantities
are total country-numbers or if they are per capita numbers. If the quantity
matrix is per capita, one should also supply a population vector for
the proper weigthing. If per capita quantities are supplied, the quantity
index returned are also per capita.
}
\concept{gk ygk pgk}
