% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Neary2004}
\alias{Neary2004}
\title{Population, Prices, and Quantities for 60 countries (1980).}
\format{
A data frame with 60 rows and 24 variables:
\describe{
\item{country}{Name of country (chr)}
\item{pop1980}{1980 population in country (in thousands) }
\item{p1}{Price of food}
\item{p2}{Price of beverages}
\item{p3}{Price of tobacco}
\item{p4}{Price of clothing and footwear}
\item{p5}{Price of gross rents}
\item{p6}{Price of fuel and power}
\item{p7}{Price of house furnishings}
\item{p8}{Price of medical care}
\item{p9}{Price of transport and communication}
\item{p10}{Price of recreation and education}
\item{p11}{Price of miscellaneous}
\item{q1}{Quantity of food (per capita)}
\item{q2}{Quantity of beverages (per capita)}
\item{q3}{Quantity of tobacco (per capita)}
\item{q4}{Quantity of clothing and footwear (per capita)}
\item{q5}{Quantity of gross rents (per capita)}
\item{q6}{Quantity of house furnishings (per capita)}
\item{q7}{Quantity of house furnishings (per capita)}
\item{q8}{Quantity of medical care (per capita)}
\item{q9}{Quantity of transport and communications (per capita)}
\item{q10}{Quantity of recreating and education (per capita)}
\item{q11}{Quantity of miscellaneous (per capita)}
}
}
\source{
Neary, J. Peter. Replication data for: Rationalizing the Penn World Table:
True Multilateral Indices for International Comparisons of Real Income.
Nashville, TN: American Economic Association (publisher), 2004.
Ann Arbor, MI: Inter-university Consortium for Political and Social Research (distributor),
2019-12-06. https://doi.org/10.3886/E116031V1
The data are licensed Creative Commons Attribution 4.0 International Public License
(https://creativecommons.org/licenses/by/4.0/).
}
\usage{
Neary2004
}
\description{
A dataset with prices and quantities for 11 goods in 60 countries 1980.
Prices and expenditure were collected around the world by the International
Comparison Project, and this is the rescaled dataset such that all US prices
are set to unity, and quantities are measured as expenditues in USD.
}
\keyword{datasets}
