% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_blocks.R
\name{p_fisher}
\alias{p_fisher}
\title{Fisher ideal price index.}
\usage{
p_fisher(P, Q)
}
\arguments{
\item{P}{An  m x n  matrix of prices (row country, column goods).}

\item{Q}{An m x n matrix of quantities (row country, column goods).}
}
\value{
An m x m matrix of country pair comparisons.
}
\description{
The Fisher ideal index takes the geometric mean of the
Paasche and the Laspeyre indices.
If there are m countries and n goods, the data are (m x n) matrices.
}
