% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{pgk}
\alias{pgk}
\title{Geary Khamis world prices}
\usage{
pgk(P, Q, pop = NULL)
}
\arguments{
\item{P}{An m x n matrix (or dataframe) of prices (row country, column goods).}

\item{Q}{An m x n matrix (or dataframe) of quantities (row country, column goods).}

\item{pop}{Optional m-vector of population numbers (if Q is per capita).}
}
\value{
An n-vector of prices.
}
\description{
This is a convenience function, which returns just the world price
vector. Calls \code{\link{gk}} and returns only the prices.
}
\examples{
P <- matrix(1:6, ncol = 2)
Q <- matrix(2:7, ncol = n)
pgk(P, Q)
}
