% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_blocks.R
\name{q_fisher}
\alias{q_fisher}
\title{Fisher ideal quantity index.}
\usage{
q_fisher(P, Q)
}
\arguments{
\item{P}{An m x n matrix of prices (row country, column goods).}

\item{Q}{An m x n matrix of quantities (row country, column goods).}
}
\value{
an m x m matrix of country pair comparisons.
}
\description{
The Fisher ideal index takes the geometric mean of the
Paasche and the Laspeyre indices.
If there are m countries and n goods, the data are (m x n) matrices.
}
