% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_blocks.R
\name{q_laspeyre}
\alias{q_laspeyre}
\title{Laspeyre quantity index.}
\usage{
q_laspeyre(pb, qt, qb)
}
\arguments{
\item{pb}{A n-vector (or m x n matrix) of prices in the baseline period.}

\item{qt}{A n-vector (or m x n matrix) of quantities in the current period.}

\item{qb}{A n-vector (or m x n matrix) of quantities in the baseline period.}
}
\value{
Price index between current an baseline period (number or vector).
}
\description{
The Laspeyre quantity index is calculated as
\deqn{ \frac{\sum_i q^t_i p^b_t}{\sum_i q^b_i p^b_i}.}
Arguments can either be vectors or matrices.
}
\examples{
Pb <- matrix(1:6, ncol = 2)
Qb <- matrix(2:7, ncol = 2)
Qt <- Qb * 1.2
p_laspeyre(Pb, Qt, Qb)
p_laspeyre(Pb[1,], Qt[1,], Qb[1,])
}
