% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{q_paasche}
\alias{q_paasche}
\title{Paasche quantity index.}
\usage{
q_paasche(pt, qt, qb)
}
\arguments{
\item{pt}{A n-vector (or m x n matrix) of prices in current period.}

\item{qt}{A n-vector (or m x n matrix) of quantities in the current period.}

\item{qb}{A n-vector (or m x n matrix) of quantities in the baseline period.}
}
\value{
Quantity index between current an baseline period (number or vector).
}
\description{
Paasche quantity index.
}
