% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_blocks.R
\name{p_paasche}
\alias{p_paasche}
\title{Paasche price index.}
\usage{
p_paasche(pt, pb, qt)
}
\arguments{
\item{pt}{A n-vector (or m x n matrix) of prices in current period.}

\item{pb}{A n-vector (or m x n matrix) of prices in the baseline period.}

\item{qt}{A n-vector (or n x n matrix) of quantities in the current period.}
}
\value{
Price index between current an baseline period (number or vector).
}
\description{
The Paasche price index is calculated as
\deqn{ \frac{\sum_i p^t_i q^t_t}{\sum_i p^b_i q^t_i}.}
Arguments can either be vectors or matrices.
}
