% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_blocks.R
\name{q_tornqvist}
\alias{q_tornqvist}
\title{Törnqvist quantity index.}
\usage{
q_tornqvist(P, Q)
}
\arguments{
\item{P}{An m x n matrix of prices (row country, column goods).}

\item{Q}{An m x n matrix of quantities (row country, column goods).}
}
\value{
an m x m matrix of country pair comparisons.
}
\description{
The Törnqvist quantity index is an alternative to the Fisher ideal index.
}
\examples{
Pt <- matrix(1:6, ncol = 2)
Qt <- matrix(2:7, ncol = 2)
q_tornqvist(Pt,Qt)
}
