% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ygk}
\alias{ygk}
\title{Geary Khamis quantity index}
\usage{
ygk(P, Q, pop = NULL)
}
\arguments{
\item{P}{An m x n matrix (or dataframe) of prices (row country, column goods).}

\item{Q}{An m x n matrix (or dataframe) of quantities (row country, column goods).}

\item{pop}{Optional m-vector of population numbers (if Q is per capita).}
}
\value{
An m-vector of quantities.
}
\description{
This is a convenience function, which returns only the quantity
index vector. Calls \code{\link{gk}} and returns only the index.
}
